requirements.txt
Django==1.7.5
djangorestframework==3.1.3
requests=2.2.1
mysql-connector-python=2.0.4
python-memcached=1.54
gunicorn=19.3.0
httpretty=0.8.10
